refactor: 🚚 consolidated config and rename sponsor svgs
Related: #175
